move out logical op runner for partition ops
